fix viewer/streamer count
